refactoring for more moes support
