+++ mlora/common/cache.py
@@ -3,8 +3,8 @@ from typing import Any, Dict, List, Optional, Tuple
 
 import torch
 
+from .config import LLMModelConfig
 from .model import Cache
-from .modelargs import LLMModelConfig
 
 
 class DynamicCache(Cache):
