+++ mlora/models/modeling_phi.py
@@ -28,7 +28,7 @@ from mlora.common import (
     flash_attention_forward,
     prepare_4d_causal_attention_mask,
 )
-from mlora.common.mix_lora import _mixtral_slice_tensor
+from mlora.common.mix_lora import _slice_tensor
 from mlora.utils import copy_parameters
 
 
@@ -339,16 +339,14 @@ class PhiMLP(LLMFeedForward):
 
             lora_name = f"moe.{moe_name}.experts.{expert_idx}"
             if lora_name in self.fc1_.loras_:
-                lora_data = _mixtral_slice_tensor(hidden_states, top_x, input_dtype)
+                lora_data = _slice_tensor(hidden_states, top_x, input_dtype)
                 act_result = act_fn(
                     self.fc1_.loras_[lora_name].forward(
-                        _mixtral_slice_tensor(common_fc1, top_x, input_dtype), lora_data
+                        _slice_tensor(common_fc1, top_x, input_dtype), lora_data
                     )
                 )
             else:
-                act_result = act_fn(
-                    _mixtral_slice_tensor(common_fc1, top_x, input_dtype)
-                )
+                act_result = act_fn(_slice_tensor(common_fc1, top_x, input_dtype))
 
             if lora_name in self.fc2_.loras_:
                 final_expert_states.append(
